iv refactoring: Split IvGL class definition into ivgl.h (from imageviewer.h).
